+++ src/network/ftp.rs
@@ -38,6 +38,7 @@ use ffi::network::ftp as ffi;
 
 /// The differents FTP modes availables.
 #[derive(Clone, PartialEq, Eq, PartialOrd, Ord, Show, Copy)]
+#[repr(i32)]
 pub enum TransferMode {
     /// Ftp Binary Mod
     FtpBinary = 0,
@@ -49,111 +50,112 @@ pub enum TransferMode {
 
 /// The status and commands id's for FTP.
 #[derive(Clone, PartialEq, Eq, PartialOrd, Ord, Show, Copy)]
+#[repr(i32)]
 pub enum Status {
     // 1xx: the requested action is being initiated,
     // expect another reply before proceeding with a new command
 
     /// Restart marker reply
-    RestartMarkerReply          = ffi::RESTARTMARKERREPLY as int,
+    RestartMarkerReply          = ffi::RESTARTMARKERREPLY as i32,
     /// Service ready in N minutes
-    ServiceReadySoon            = ffi::SERVICEREADYSOON as int,
+    ServiceReadySoon            = ffi::SERVICEREADYSOON as i32,
     /// Data connection already opened, transfer starting
-    DataConnectionAlreadyOpened = ffi::DATACONNECTIONALREADYOPENED as int,
+    DataConnectionAlreadyOpened = ffi::DATACONNECTIONALREADYOPENED as i32,
     /// File status ok, about to open data connection
-    OpeningDataConnection       = ffi::OPENINGDATACONNECTION as int,
+    OpeningDataConnection       = ffi::OPENINGDATACONNECTION as i32,
 
     // 2xx: the requested action has been successfully completed
 
     /// Command ok
-    Ok                          = ffi::OK as int,
+    Ok                          = ffi::OK as i32,
     /// Command not implemented
-    PointlessCommand            = ffi::POINTLESSCOMMAND as int,
+    PointlessCommand            = ffi::POINTLESSCOMMAND as i32,
     /// System status, or system help reply
-    SystemStatus                = ffi::SYSTEMSTATUS as int,
+    SystemStatus                = ffi::SYSTEMSTATUS as i32,
     /// Directory status
-    DirectoryStatus             = ffi::DIRECTORYSTATUS as int,
+    DirectoryStatus             = ffi::DIRECTORYSTATUS as i32,
     /// File status
-    FileStatus                  = ffi::FILESTATUS as int,
+    FileStatus                  = ffi::FILESTATUS as i32,
     /// Help message
-    HelpMessage                 = ffi::HELPMESSAGE as int,
+    HelpMessage                 = ffi::HELPMESSAGE as i32,
     /// NAME system type, where NAME is an official system name from the list in the Assigned Numbers document
-    SystemType                  = ffi::SYSTEMTYPE as int,
+    SystemType                  = ffi::SYSTEMTYPE as i32,
     /// Service ready for new user
-    ServiceReady                = ffi::SERVICEREADY as int,
+    ServiceReady                = ffi::SERVICEREADY as i32,
     /// Service closing control connection
-    ClosingConnection           = ffi::CLOSINGCONNECTION as int,
+    ClosingConnection           = ffi::CLOSINGCONNECTION as i32,
     /// Data connection open, no transfer in progress
-    DataConnectionOpened        = ffi::DATACONNECTIONOPENED as int,
+    DataConnectionOpened        = ffi::DATACONNECTIONOPENED as i32,
     /// Closing data connection, requested file action successful
-    ClosingDataConnection       = ffi::CLOSINGDATACONNECTION as int,
+    ClosingDataConnection       = ffi::CLOSINGDATACONNECTION as i32,
     /// Entering passive mode
-    EnteringPassiveMode         = ffi::ENTERINGPASSIVEMODE as int,
+    EnteringPassiveMode         = ffi::ENTERINGPASSIVEMODE as i32,
     /// User logged in, proceed. Logged out if appropriate
-    LoggedIn                    = ffi::LOGGEDIN as int,
+    LoggedIn                    = ffi::LOGGEDIN as i32,
     /// Requested file action ok
-    FileActionOk                = ffi::FILEACTIONOK as int,
+    FileActionOk                = ffi::FILEACTIONOK as i32,
     /// PATHNAME created
-    DirectoryOk                 = ffi::DIRECTORYOK as int,
+    DirectoryOk                 = ffi::DIRECTORYOK as i32,
 
     // 3xx: the command has been accepted, but the requested action
     // is dormant, pending receipt of further information
     /// User name ok, need password
-    NeedPassword                = ffi::NEEDPASSWORD as int,
+    NeedPassword                = ffi::NEEDPASSWORD as i32,
     /// Need account for login
-    NeedAccountToLogIn          = ffi::NEEDACCOUNTTOLOGIN as int,
+    NeedAccountToLogIn          = ffi::NEEDACCOUNTTOLOGIN as i32,
     /// Requested file action pending further information
-    NeedInformation             = ffi::NEEDINFORMATION as int,
+    NeedInformation             = ffi::NEEDINFORMATION as i32,
 
     // 4xx: the command was not accepted and the requested action did not take place,
     // but the error condition is temporary and the action may be requested again
 
     /// Service not available, closing control connection
-    ServiceUnavailable          = ffi::SERVICEUNAVAILABLE as int,
+    ServiceUnavailable          = ffi::SERVICEUNAVAILABLE as i32,
     /// Can't open data connection
-    DataConnectionUnavailable   = ffi::DATACONNECTIONUNAVAILABLE as int,
+    DataConnectionUnavailable   = ffi::DATACONNECTIONUNAVAILABLE as i32,
     /// Connection closed, transfer aborted
-    TransferAborted             = ffi::TRANSFERABORTED as int,
+    TransferAborted             = ffi::TRANSFERABORTED as i32,
     /// Requested file action not taken
-    FileActionAborted           = ffi::FILEACTIONABORTED as int,
+    FileActionAborted           = ffi::FILEACTIONABORTED as i32,
     /// Requested action aborted, local error in processing
-    LocalError                  = ffi::LOCALERROR as int,
+    LocalError                  = ffi::LOCALERROR as i32,
     /// Requested action not taken; insufficient storage space in system, file unavailable
-    InsufficientStorageSpace    = ffi::INSUFFICIENTSTORAGESPACE as int,
+    InsufficientStorageSpace    = ffi::INSUFFICIENTSTORAGESPACE as i32,
 
     // 5xx: the command was not accepted and
     // the requested action did not take place
     /// Syntax error, command unrecognized
-    CommandUnknown              = ffi::COMMANDUNKNOWN as int,
+    CommandUnknown              = ffi::COMMANDUNKNOWN as i32,
     /// Syntax error in parameters or arguments
-    ParametersUnknown           = ffi::PARAMETERSUNKNOWN as int,
+    ParametersUnknown           = ffi::PARAMETERSUNKNOWN as i32,
     /// Command not implemented
-    CommandNotImplemented       = ffi::COMMANDNOTIMPLEMENTED as int,
+    CommandNotImplemented       = ffi::COMMANDNOTIMPLEMENTED as i32,
     /// Bad sequence of commands
-    BadCommandSequence          = ffi::BADCOMMANDSEQUENCE as int,
+    BadCommandSequence          = ffi::BADCOMMANDSEQUENCE as i32,
     /// Command not implemented for that parameter
-    ParameterNotImplemented     = ffi::PARAMETERNOTIMPLEMENTED as int,
+    ParameterNotImplemented     = ffi::PARAMETERNOTIMPLEMENTED as i32,
     /// Not logged in
-    NotLoggedIn                 = ffi::NOTLOGGEDIN as int,
+    NotLoggedIn                 = ffi::NOTLOGGEDIN as i32,
     /// Need account for storing files
-    NeedAccountToStore          = ffi::NEEDACCOUNTTOSTORE as int,
+    NeedAccountToStore          = ffi::NEEDACCOUNTTOSTORE as i32,
     /// Requested action not taken, file unavailable
-    FileUnavailable             = ffi::FILEUNAVAILABLE as int,
+    FileUnavailable             = ffi::FILEUNAVAILABLE as i32,
     /// Requested action aborted, page type unknown
-    PageTypeUnknown             = ffi::PAGETYPEUNKNOWN as int,
+    PageTypeUnknown             = ffi::PAGETYPEUNKNOWN as i32,
     /// Requested file action aborted, exceeded storage allocation
-    NotEnoughMemory             = ffi::NOTENOUGHMEMORY as int,
+    NotEnoughMemory             = ffi::NOTENOUGHMEMORY as i32,
     /// Requested action not taken, file name not allowed
-    FilenameNotAllowed          = ffi::FILENAMENOTALLOWED as int,
+    FilenameNotAllowed          = ffi::FILENAMENOTALLOWED as i32,
 
     // 10xx: SFML custom codes
     /// Response is not a valid FTP one
-    InvalidResponse             = ffi::INVALIDRESPONSE as int,
+    InvalidResponse             = ffi::INVALIDRESPONSE as i32,
     /// Connection with server failed
-    ConnectionFailed            = ffi::CONNECTIONFAILED as int,
+    ConnectionFailed            = ffi::CONNECTIONFAILED as i32,
     /// Connection with server closed
-    ConnectionClosed            = ffi::CONNECTIONCLOSED as int,
+    ConnectionClosed            = ffi::CONNECTIONCLOSED as i32,
     /// Invalid file to upload / download
-    InvalidFile                 = ffi::INVALIDFILE as int
+    InvalidFile                 = ffi::INVALIDFILE as i32
 }
 
 /// The FTP client
@@ -199,7 +201,7 @@ impl ListingResponse {
     /// Return the status code
     pub fn get_status(&self) -> Status {
         unsafe {
-            mem::transmute(ffi::sfFtpListingResponse_getStatus(self.listing_response) as i16)
+            mem::transmute(ffi::sfFtpListingResponse_getStatus(self.listing_response) as i32)
         }
     }
 
@@ -263,7 +265,7 @@ impl DirectoryResponse {
     /// Return the status code
     pub fn get_status(&self) -> Status {
         unsafe {
-            mem::transmute(ffi::sfFtpDirectoryResponse_getStatus(self.directory_response) as i16)
+            mem::transmute(ffi::sfFtpDirectoryResponse_getStatus(self.directory_response) as i32)
         }
     }
 
@@ -315,7 +317,7 @@ impl Response {
     /// Return Status code
     pub fn get_status(&self) -> Status {
         unsafe {
-            mem::transmute(ffi::sfFtpResponse_getStatus(self.response) as i16)
+            mem::transmute(ffi::sfFtpResponse_getStatus(self.response) as i32)
         }
     }
